+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Added "Mechanized Lumber Extractor" to herbalism mounts list.
+* Added option "Try to copy target's mount".